chore: renaming and cleaning up small bits
