+++ Tests/AuthenticationTests/Mocks/Helpers/MockOIDAuthorizationService_Success.swift
@@ -8,7 +8,7 @@ class MockOIDAuthorizationService_Success: OIDAuthorizationService {
         prefersEphemeralSession: Bool,
         callback: @escaping OIDAuthorizationCallback
     ) -> any OIDExternalUserAgentSession {
-        let session = MockOIDExternalUserAgentSession()
+        let session = MockOIDExternalUserAgentSession_Success()
         session.callback = callback
         return session
     }
