+++ Tests/AuthenticationTests/Mocks/Helpers/MockOIDExternalUserAgentSession_False.swift
@@ -1,0 +1,16 @@
+import AppAuthCore
+
+class MockOIDExternalUserAgentSession_False: NSObject,
+                                             OIDExternalUserAgentSession {
+    var callback: OIDAuthorizationCallback?
+    
+    public func cancel() { }
+    
+    public func cancel() async { }
+    
+    public func failExternalUserAgentFlowWithError(_ error: Error) { }
+    
+    public func resumeExternalUserAgentFlow(with URL: URL) -> Bool {
+        return false
+    }
+}
