+++ Tests/AuthenticationTests/Mocks/TokenErrors/TokenInvalidClient/MockOIDAuthorizationService_TokenInvalidClientError.swift
@@ -9,7 +9,7 @@ class MockOIDAuthorizationService_TokenInvalidClientError: OIDAuthorizationServi
         prefersEphemeralSession: Bool,
         callback: @escaping OIDAuthorizationCallback
     ) -> any OIDExternalUserAgentSession {
-        let session = MockOIDExternalUserAgentSession()
+        let session = MockOIDExternalUserAgentSession_Success()
         session.callback = callback
         return session
     }
