+++ Tests/AuthenticationTests/Mocks/TokenErrors/TokenInvalidGrant/MockOIDAuthorizationService_TokenInvalidGrantError.swift
@@ -8,7 +8,7 @@ class MockOIDAuthorizationService_TokenInvalidGrantError: OIDAuthorizationServic
         prefersEphemeralSession: Bool,
         callback: @escaping OIDAuthorizationCallback
     ) -> any OIDExternalUserAgentSession {
-        let session = MockOIDExternalUserAgentSession()
+        let session = MockOIDExternalUserAgentSession_Success()
         session.callback = callback
         return session
     }
