+++ Tests/AuthenticationTests/Mocks/TokenErrors/TokenInvalidRequest/MockOIDAuthorizationService_TokenInvalidRequestError.swift
@@ -9,7 +9,7 @@ class MockOIDAuthorizationService_TokenInvalidRequestError: OIDAuthorizationServ
         prefersEphemeralSession: Bool,
         callback: @escaping OIDAuthorizationCallback
     ) -> any OIDExternalUserAgentSession {
-        let session = MockOIDExternalUserAgentSession()
+        let session = MockOIDExternalUserAgentSession_Success()
         session.callback = callback
         return session
     }
