+++ Tests/AuthenticationTests/Mocks/TokenErrors/TokenUnauthorizedClient/MockOIDAuthorizationService_TokenUnauthorizedClientError.swift
@@ -9,7 +9,7 @@ class MockOIDAuthorizationService_TokenUnauthorizedClientError: OIDAuthorization
         prefersEphemeralSession: Bool,
         callback: @escaping OIDAuthorizationCallback
     ) -> any OIDExternalUserAgentSession {
-        let session = MockOIDExternalUserAgentSession()
+        let session = MockOIDExternalUserAgentSession_Success()
         session.callback = callback
         return session
     }
