+++ Tests/AuthenticationTests/Mocks/TokenErrors/TokenUnknownError/MockOIDAuthorizationService_TokenUnknownError.swift
@@ -8,7 +8,7 @@ class MockOIDAuthorizationService_TokenUnknownError: OIDAuthorizationService {
         prefersEphemeralSession: Bool,
         callback: @escaping OIDAuthorizationCallback
     ) -> any OIDExternalUserAgentSession {
-        let session = MockOIDExternalUserAgentSession()
+        let session = MockOIDExternalUserAgentSession_Success()
         session.callback = callback
         return session
     }
