+++ Tests/AuthenticationTests/Mocks/TokenErrors/TokenUnsupportedGrantType/MockOIDAuthorizationService_TokenUnsupportedGrantTypeError.swift
@@ -9,7 +9,7 @@ class MockOIDAuthorizationService_TokenUnsupportedGrantTypeError: OIDAuthorizati
         prefersEphemeralSession: Bool,
         callback: @escaping OIDAuthorizationCallback
     ) -> any OIDExternalUserAgentSession {
-        let session = MockOIDExternalUserAgentSession()
+        let session = MockOIDExternalUserAgentSession_Success()
         session.callback = callback
         return session
     }
